api for range_topic, score_scale

+++ src/main/java/com/doan2025/webtoeic/controller/RangeTopicController.java
@@ -21,7 +21,7 @@ public class RangeTopicController {
     private final RangeTopicService rangeTopicService;
 
     @GetMapping()
-    public ApiResponse<?> getRangeTopic(HttpServletRequest request, Integer id) {
+    public ApiResponse<?> getRangeTopic(HttpServletRequest request, Long id) {
         return ApiResponse.of(
                 ResponseCode.GET_SUCCESS,
                 ResponseObject.RANGE_TOPIC,
